Project Refactoring
Removed all the lua code, swift is now going to be the backbone of the
project. I am updating modules today and then pushing them to the
project as they become stable. Right now, the project should be
downloadable and then you go look at the views on your phone/computer.